Move parsing logic to a separate file
