+++ src/types.ts
@@ -1,0 +1,8 @@
+export interface Color {
+  red: number;
+  green: number;
+  blue: number;
+  alpha: number;
+}
+
+export type Theme = Map<string, Color>;
